requirements.txt
Flask
langchain
pytest
python-dotenv
psycopg
weaviate-client
aim
Flask-Session
pydantic
